Force Content-Type for localized file extensions
I thought I had nailed this out in #21 with the newer Apache and new
OpenShift config but I was wrong and could not get it to work again, I
must have made a mistake.

Instead of just reinstating the previous list of AddHandler calls, let's
generate it here in order to avoid some tedious maintenance.